+++ packages/plugin-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @fusion/plugin-sdk
 
+## 0.73.0-beta.3
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.3
+
+## 0.73.0-beta.2
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.2
+
+## 0.73.0-beta.1
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.1
+
+## 0.73.0-beta.0
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.0
+
 ## 0.72.0
 
 ### Patch Changes