+++ plugins/fusion-plugin-grok-runtime/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @fusion-plugin-examples/grok-runtime
 
+## 0.2.9-beta.3
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.3
+- @fusion/plugin-sdk@0.73.0-beta.3
+
+## 0.2.9-beta.2
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.2
+- @fusion/plugin-sdk@0.73.0-beta.2
+
+## 0.2.9-beta.1
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.1
+- @fusion/plugin-sdk@0.73.0-beta.1
+
+## 0.2.9-beta.0
+
+### Patch Changes
+
+- @fusion/core@0.73.0-beta.0
+- @fusion/plugin-sdk@0.73.0-beta.0
+
 ## 0.2.8
 
 ### Patch Changes